ImmoScout24 Integration weiter bearbeitet. Ab jetzt kann per URL, definiert im HouseController eine Suche per Stadt (Parameter city) und Radius (Parameter radius) probiert werden. Es wird eine JSON Struktur von House Instanzen zurück gegeben.

+++ src/main/java/de/immoPiraten/realEstate/House.java
@@ -5,13 +5,13 @@ import javax.persistence.Entity;
 @Entity
 public class House extends ResidentialRealEstate {
 
-	private int landArea;
+	private double landArea;
 
-	public int getLandArea() {
+	public double getLandArea() {
 		return landArea;
 	}
 
-	public void setLandArea(int landArea) {
+	public void setLandArea(double landArea) {
 		this.landArea = landArea;
 	}
 }
